ml-services/chatbot/requirements.txt
streamlit
langchain
langchain-core
langchain-groq
python-dotenv
httpx
pydantic
fastapi
uvicorn